Officially finished the export command.
I said I was done with it yesterday, but it was only a simple PoC with a lot of missing features (related to flag comprehension), a bug that I discovered today (when using the --to flag - which I didn't test yesterday becasue even the --to flag wasn't supported), unhandled errors, a bunch of trash-code laying around, and broken docs. Now, it's all officially DONE. Will not be touched again for a while, probably (maybe only to add a longer description).

The catch is that an underlying library, trans.Export, still throws a lot of „not-implemented” errors. But that's TECHNICALLY a separate thing, and also the command now properly handles them, at least (which wasn't the case the last time).