Made it so that server-only mods will be disabled on the client in Dev Mode
I thought this won't be necessary for modern Minecraft (where mods usually just don't load if sides don't match), but Create Track Map proved me wrong. I mean... It's TECHNICALLY it's not a server-only mod (even if GhostLand treats it as such), but it has a nasty habit of crashing the client, so we not only have to make sure it won't load on client prod exports (as it already is, by the virtue of being marked as server-side only, ie. Modrinth will simply ignore it), but in dev as well.

Also, to make it so that previosly-disabled mods get re-enabled, we had to add a new system: prefix-driven disabled status. If Grinch's prefixes (there are now two) are present, disabled status will be completley ignored, instead being derived from said prefix. Since this is a pretty massive change, we decided to bump config versions (even though it's TECHNICALLY not for that - but, unfortunatley, I lacked the foresight to make a versioning system for prefixes), so people won't accidentially enable mods that were supposed to stay disabled.

Btw, I didn't test this at all, as I needed these changes in a pinch. If something's broken - I'm seconds away from finding out, anyway.